+++ frontend/src/components/Inputs/Password.vue
@@ -18,7 +18,7 @@
         </p-button>
       </span>
     </div>
-    <div class="validate" v-if="!valid_password && !this.$store.state.app.login">             
+    <div class="validate" v-if="validate && !valid_password">             
       <p v-if="!this.contains_eight_characters"> 8 Characters Long, </p> 
       <p v-if="!this.contains_number"> Number,</p> 
       <p v-if="!this.contains_uppercase"> Uppercase, </p> 
@@ -35,7 +35,8 @@
     props: {
       value: String,
       label: String,
-      placeholder: String
+      placeholder: String,
+      validate: Boolean
     },
     data: () => ({
       valid_password: false,
